Hide darktable.image_cache implementation details
1. We don't want/need to expose the usage of `darktable.image_cache' in all functions using the image_cache
   for denser code and readability.
2. Allocation and free image_cache is done inside the init/cleanup functions
3. Some additional safety checks for dereferencing making sure operations are done only with a valid dt_image_t
4. Some leading underscore renaming for static functions
5. Some asserts added for debugging builds

+++ src/control/jobs/image_jobs.c
@@ -36,7 +36,7 @@ static int32_t dt_image_load_job_run(dt_job_t *job)
 
   if(buf.buf && buf.height && buf.width)
   {
-    const double aspect_ratio = (double)buf.width / (double)buf.height;
+    const float aspect_ratio = (float)buf.width / (float)buf.height;
     dt_image_set_aspect_ratio_if_different(params->imgid, aspect_ratio, FALSE);
   }
